fix: keep embedded fixture loader within execution package
